requirements.txt
openai
requests
numpy
ollama
Pillow